backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langgraph>=0.0.20
faiss-cpu>=1.7.4
pydantic>=2.5.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
openai>=1.10.0
tiktoken>=0.5.2